Optimize grep command search to avoid calling executable-find too much
* Makefile: Improve compilation command
* test/unit-tests.el: Use the new grep api.
* counsel-gtags.el: Move grep search api to a different file. This makes
the code a bit modular and readable.
(counsel-gtags--grep-commands-list) : Removed as the info is redundant
in counsel-gtags-grep-command-options-alist.
(counsel-gtags--grep-command-find) : new function trying to optimize the
executable-find calls. Uses buffer-local, global variable and
buffer-local caching.

For local files it shouldn't be needed to call executable find and for
remote ones only one time should be enough.